+++ src/Models/asset/Asset.ts
@@ -1,0 +1,7 @@
+import AssetInterface from '../../Interfaces/asset/Asset';
+
+export default class Asset {
+    constructor(parameters: AssetInterface) {
+        Object.assign(this, parameters);
+    }
+}