+++ src/Models/asset/AssetSource.ts
@@ -1,0 +1,7 @@
+import AssetSourceInterface from '../../Interfaces/asset/AssetSource';
+
+export default class AssetSource {
+    constructor(parameters: AssetSourceInterface) {
+        Object.assign(this, parameters);
+    }
+}